Updates
- Remove preference pane requirement. The test is not long enough for that. To be too far off could be quite disheartening for the interviewee, although we say we don't expect them to finish.
- Remove specific clone command. It pointed to the incorrect repo and any method of cloning is fine.
- Improve the securities quotes api documentation
    - link to symbols list to help make sure it's not missed
    - mention that they can be comma separated in the symbols query parameter
    - put an example to show the data structure of the response

We discussed a section to say we're not so interested in a specific framework, but one may be used. I couldn't find a way to express our thinking in a way that wasn't off-putting. It might be better to leave to the interviewer to ask questions about their openness to other frameworks/vanilla js, or prompt to drop it if there's any unexpected integration/build issues that take up too much time?